src/requirements.txt

transformers==2.3.0
numpy==1.18.1 
torchvision==0.5.0
tensorboardX